Fixed graphical artifacts appearing in Fake Map Border/Off Map terrains with SDL 2.0.6+ (fixes #2061)
[ci skip]

Something was screwy about these image's alpha. Normally when I remove an alpha channel in GIMP,
it shows up white. With these images, it showed up black. Maybe they had an embedded palette or
something, I don't know, but the problem is fixed.

Something must have changed in SDL 2.0.6 that caused this issue to appear. Not sure if anything
needs to be patched in our source code, though (but not likely)