GUI2/Window: don't constantly send useless DRAW events to the queue
This is a small subset of a change made on accelerated_rendering. Essentially, since drawing
doesn't happen until events::pump is called, and that function also drops all but one DRAW
event from the queue, sending a DRAW event every 20 ms is just useless noise and excessive
CPU use.

There should be no change in the actual rate GUI2 dialogs are redrawn with this change.