experiment to fix edge-of-hex artifacts with fractional zooming
The idea is, when scaling a hex to zoom, instead of masking it
first and then scaling it up, it should be scaled to zoom and
then masked with a scaled up terrain mask image, because it will
be more precise.

I find that if this is done, and scale to hex algorithm is set to
nearest neighbor, then I no longer get any artifacts from this.

This is an adaptation of #318 by cbeck88