Create a save_index entry when saving games
Prior to this, if you saved a game with a campaign-specific leader sprite and exited the game
or reloaded the cache *without* opening the Load Game dialog, the leader image would appear as
Unknown. This is because the save_index entry (and the accompanying path expansion) only occurred
when the save was selected in the dialog.

This change ensures all saves' leaders' images from here on out will always appear in the Load Game
dialog (unless, of course, the save_index file is deleted).