Fix #2067: Scrolling position in Lua console is reset constantly
The problem is that a line that's too wide to fit into the window
invalidates the layout of the entire window, and the window will lay itself
out the next time it will be drawn, after the Lua console code has scrolled
it to bottom.

The easiest fix is to simply defer scrolling to bottom until the window has
been drawn (and thus laid out) next time.