Wesnothd: move some function declarations to more logical places
send_to_players was only used in the game class implementation. The other two functions were
used in both server.cpp and game.cpp, so it made sense to be declared in server.hpp.

For some damn reason, I needed to specify `game::error` in the global namespace to get this
to build now. No idea why.