Fix: canceling attempt to reconnect didn't close addon manager
Only happened if the player used the buttons in the addon list: the buttons
in the right panel worked fine. The problem was that the addon list caught
the addons_client::user_exit exception separately and didn't rethrow it
when necessary, unlike the addon manager itself.

I decided to simply throw a different exception instead.