Revert part of "GUI2 Textbox: Remove the other two emacs keybindings"
This reverts part of commit 468c6e0f494793d57b0b1264894cd7f0b0383fa2.

Ctrl+A was removed because its emacs meaning is surprising to some
people.  Ctrl+E was removed because it was the counterpart of Ctrl+A.
Ctrl+U however doesn't conflict with anything else, and it's useful
to some people (me), so reinstate it.