fix crash on #wesnoth.unit_types
`#wesnoth.unit_types` should simply return the number of unit types.

This fixes the crash and reports the correct value: the number of unit types defined.

This does not fix the rest of the crap like pairs() crawls, you can't interate variations or get the variation count.

My knee-jerk reaction is to delete all scripting support for unit_types so that it's filed where it belongs: in the Trash. But, what the hell, nobody has complained about what utter shit this code is, so I'll just fix this so we don't crash and carry on working on getting the code to actually do what it's supposed to instead of the lame-ass shit it currently does.

Can you tell what I think of this file?

closes #2381