Renames the Dunefolk(formerly Khalifate) units based on forum discussions.
See this thread, along with a few linked within:
https://forums.wesnoth.org/viewtopic.php?f=12&t=47704

The purpose of this is both to:
a) Make the unit names more descriptive, so new players especially will have an easier time learning them and knowing what they do.
b) Make them eligible for translation, like the units of other factions are.

This commit only renames the files, along with the faction/AoH faction and wmllint.