Fix C++ unit tests
Regression from commit f2b31ba082e1ff47443120469e06b89318f37284.

The calls to get_fake_display() must not be removed. They are essential to
change the size of the fake screen. Fortunately we have a unit test to
check that constructing GUI2 windows doesn't succeed if the screen is way
too small.