Enables travis to cache scons' build/ directory.
Scons makes this simple - all the object files are placed in build/release/, and the build/sconsign.dblite file keeps track of the md5 sum of each file built. This then means all that's necessary to prevent scons from doing a full rebuild each time is to cache the build/ directory.

Keep in mind however, that this does not make the actual compiling faster, it just means that most of the time fewer things will need to be compiled. Any change that would trigger a rebuild of most or all of the source files will still take just as long as before.