MP: add alot of xxx_lock setting
In these cases they have only in limited situations an effect:
In normal play the sides are hidden, if not using map settings they are ignored,
so this mostly only has an effect if starting the game in debug mode an using map
settings.

Other ones revert acidental changes caused by force_lock_settings.
fore_lock_settings does not only disallow the player to modify the settings,
but does also serve as default value for all the other lock settings.