Quit pretending that underscores in savefile names are actually spaces
The code was frantically converting spaces to underscores and back to keep
the facade up, and sometimes it failed (#1567).

The solution is obvious: simply use spaces for real in savefile names.
This way the conversion is only needed in one place (when generating the
suggested filename).

Fixes #1567. As a minor bonus, this also makes the "save game" dialog show
the filename with spaces.