campaignd: Reject empty passphrases
Fixes issue #2444.

Clients may send an empty passphrase, especially if they do not do as
the mainline client and generate a random one to substitute in the .pbl.
An empty string is still hashable, but the server is really not supposed
to allow an empty passphrase so we should just reject those. It's also a
good warning sign against broken clients -- like wesnoth_addon_manager
right now.