Fix bad design with a slimy hack
AKA Check the fucking Travis-CI logs before you merge, damnit!

GCC wanted a function prototype becore declaration. This is noise: added a proto to silence it.

We do NOT want to drag the translations into the servers. So have each define a do-nothing implemenation of the deprecation function.

I don't CARE if this is gross and disgusting. It's the cleanest way to get us compiling again. It's only a few hours until we tag a release and I don't want to see yet another slip in the schedule due to a stupid mistake.