Fixes cmake's cache getting corrupted in certain situations.
This adds functionality to check if a failed build used cmake, and if it did, clear out everything that was compiled during the failed run so that they will be treated as new and guaranteed to be rebuilt.

It also adds a inverted grep match on make to filter out the thousands of "is newer than" messages that can occasionally appear and clog up the log.