applied fix that pofix.py already applies
A fix to httt causes fuzzy strings in NR without this change. Tow
options were available:
1) Remove the fix from pofix since it creates false matches
2) Add the fix to the file in NR that also has the issue and apply
pofix to all po files in wesnoth-nr textdomain as well as the pot file

Since it is a "bug" having this grammar mistake in there, I went for
option 2 even though it is string freeze (since it should no longer
cause "false positive fuzzies").