Fix missing explicit keyword
cppcheck (via Codacy) reports that a member function with one parameter should use the explicit keyword.

In this case, the error pointed out that we were missing an assignment operator to match the constructor.

Boost gets confused by all this so, in a few places, explicitly construct a temporary object.

Also, it's more clear the use the str() function instead of depending upon implicit conversion to std:string.