Promote ai_info_ from pointer to a member object
We used to create a new ai_info object in the constructor of
playsingle_controller in order to reset it on scenario changes.
Now when the AI manager is a part of game state, resetting the game state
implicitly creates a new ai_info object, and setting the AI info manually
is no longer necessary.

This also ensures that ai_info can't ever be null, fixing #2522.

In the process, I also discovered that ai::registry::init() doesn't do
anything, so I removed it, its declaration, and ai/registry.hpp which
contained nothing except said function declaration.