Fix AppVeyor for VS 2017
The idiots at Microsoft, in their infinte wisdom, have suddenly decided that if it looks like you have a CMake configuration in your source tree that you want to use CMake instead of MSBuild and automatically start up a background CMake and makefile run.

This, of course, means that it fails because CMake can not find the external dependencies.

It also means AppVeyor hangs until we timeout because, as a background process, AppVeyor never sees the job is running. To make matters worse, this seems to actually try to compile the full source, at a glacially slow rate.

And, with all due praises to their godly infalibility, Microsoft has not provided any means to stop using CMake.

So, while it's not a perfect upgrade, so there may be minor differences from what you'll see using the full IDE interactively, we can override the PlatformToolset to use v141 while using MSBuild with the Visual Studio 2013 project files.

All Hail Microsoft!