Play_controller: register as AI observer in reset_gamestate()
Now when play controller can outlive AI manager, it needs to register
itself as an observer when the AI manager is recreated.

While I was working with this code, I also moved the existing
registration from playsingle_controller to play_controller (for
consistency) and stopped using the singleton pointer to access the
manager.

Thanks to @gfgtdf for spotting the problem.