Editor: fixed village counter possibly being incorrect after resizing map (fixes #1390)
On looking into this, it turns out the [replace_map] WML action totally clears the village vector
and repopulates it from the new map. That's why it was reported didn't affect that tag. The editor,
OTHO, actually resizes the map, not recreates it.

I decided to use @wedge009 's solution described in the bug above and manually remove off-map
villages from the map's village vector. Something must also have changed in the past two years
since he attempted his fix, since the theme counter now immediately displays the correct value.

As for the issue he raised about villages on the map border, I they're simply ignored. Villages
placed there when you're drawing a map are not counted anyway, and with this change they are
correctly removed from the village count.