LoW S14 Objectives, reinforcements, edit repetitive dialog (ok for string freeze)
Objectives: note that the death of Galtrid is a losing condition

Reinforcements: In multiplayer, Landar's L3 troops that join on turn 9
now join his side, instead of trying to join side 13 (and ending up on
Kalenz' side).

Dialog: the following conversation happened on both turn 1 and turn 9.
Having 6 repeated messages in a row gave a strong sense of deja vu,
and I thought that a bug had repeated an event. This commit cuts the
repetition.

* Elf: "Kalenz! We come to fight beside you!"
* Kalenz: "Where is the rest of the elvish army? ...
* Elf: "The Great Council has decided it was too risky ...
* Landar: "That is well! If the Ka’lian is too fearful or blind ...
* Cleodil: "It is not well that we have become so divided as this."
* Kalenz: "No, it is not. But ...