campaignd: Add support for free-form tags on add-ons
This adds a tags= attribute to the pbl WML that can be used to specify
comma-separated tags to let users filter on them on the Add-ons Manager.

We don't do anything fancy with the tags at the moment since they can be
matched with the existing filter algorithm without any changes, with the
caveat that the word separator the filter box uses is whitespace instead
of commas -- meaning that to match multiple tags they must be specified
with spaces between them, in any order, and that pasting the tags from
the WML into the filter box will not yield a match as some people might
expect.

There is also no UI for the tags right now other than the filter box, so
they are little more than an implementation detail hidden from the user.

See also issue #2565.