MP Join Game: removed unnecessary event context when prompting faction selection
This was a holdover from the old GUI1 dialog code, IIRC. It's not needed with GUI2, and is
probably the reason I had that weird issue where I couldn't call this function (fetch_game_config)
from mp_join_game::pre_show. However, even if that were the case, the current design is
perfectly acceptable.