Help: disable Unknown Topic popup (fixes #2587)
It's just a workaround and doesn't address the underlying issue, yes, but given
that we intend to refactor the help browser in 1.15, it's not worth the time and
effort to solve the event context layering issue, at least not for this.

The underlying issue might need fixing anyway, but I'll leave that for another time.