Game Events: don't write disabled events to config
This should help mitigate an odd issue some people have reported with 1.13.11. After reloading a
save *after* executing a first_time_only=yes event, they can execute the event again. Looking at
savefiles for one of the reports, the event was included in the savefile even though it had already
executed. Most likely, when the player reloaded, the event was re-added to the gamestate and they
could execute it again.

The questions is then why the event had never been removed in handler cleanup. I'll have to look
into that further, but this at least ensures we don't get dead events written to config (that
includes saved games).