fix no_leader having no effect (#2606)
fix no_leader having no effect

fixes #2604

it previously only worked if both leader_lock and no_leader were set to yes. Now it the type= is preserved if any of the two are given, the difference bewteen those two is that no_leader only works is type= is empty. 

this also fixes some wrong parameternames in the the flg_manager ctor declaration.

This commit does not change the default value for those keys: no_leader default to yes in [side]s in [scenario] but not in multiplayer. While leader_lock defaults to the force_lock_settings key in [scenario]/[multiuplayer]