Preferences Dialog: don't use invalid initial selection for combo advanced prefs
Fixes #2612. Bug above was caused with a legacy value for compress_saves ("no") in the user's
preferences file. Since that didn't match an option ID, the game passed an invalid value to
menu_button::set_values, which then asserted. It's not really worth it to add value aliases to
handle the legacy compression values, so we just reset the initial selection to 0 if an invalid
(in this case, one of the legacy values) is used. Might mean some user's preferences get
inadvertently reset,but they can simply update their selections again. At least it's not crashing.