Change more mentions of "Shadow Master" to "shadowm"
The one that needs a two-word name changes to Shikadi_Queen instead.

Was code-reviewing LoW, and spotted the comment for the RANDOM_PLACEMENT macro.
I also checked if that could be replaced with the core SCATTER_UNITS macro, but
SCATTER_UNITS seems to be overcomplex for the case of placing a single unit.