LoW save the persistent multiplayer data for chapter 2 (fixes #2637)
In multiplayer mode, the LoW campaign has multiple start points, and some
persistent storage that stores the characters at the end of a chapter, so that
a new game starting from the next chapter starts with those stats.

The data for chapter 2 wasn't saved, because the campaign tried to create it in
a cutscene scenario that doesn't have the expected sides.