GUI2/Widget: removed dialog() getter
The doc comment indicated this should be removed once dialog_callback was refactored out,
which has now been done.

This function just a wrapper for window::dialog(), which I left. That makes sense, since
windows and dialogs are closely related. If one *really* still needs to get the dialog in
which a widget is placed, they can still use widget.get_window().dialog().