GUI2/Modal Dialog, Modeless Dialog: made window_ a protected class member
This will allow direct use of the window pointer without binding in hundreds of different
callbacks just because we want to use it for, say, find_widget.

This is just the base change. I still might change my mind as to this method, and I have
yet to undertake any refactoring around this.

modal_dialog::get_window and modeless_dialog::get_window remain.