Convert the rest of the changelog to Markdown
Sort of.

The only things that are converted here are the version headings and the
category sub-headings, up to and including version 1.1.9. Versions older
than that have mixed root items and categories, and converting those to
Markdown that renders correctly on GitHub without mixing up extraneous
root items with the categories' own entries requires some non-trivial
hand-fixing that I'm not going to do (at least not right now).

[ci skip]