Save screenshots as JPGs when building with SDL_Image 2.0.2 or newer
This functionality was added in that release. Since it's only 4 months old, it's
too early to make minimum; it won't be available on some Linux distros yet.

There's currently no way to choose between JPG or PNG files. That can (and should)
be added later. For now, it defaults to JPG files if SDL_Image is of the appropriate
version.

I've also left the BMP fallback path, but that should never be reached now, since
PNG saving is always available

Closes #1448.