Adds a flag to globally enable or disable unit tests.
This flag is set from the Travis settings page, so in the future this value can be flipped on or off without needing to make a commit.  This is mostly motivated by the fact that running the unit tests on master at the moment is fairly pointless, and disabling unit tests until fixing them is a priority will allow the builds to succeed if there are at least no compile errors rather than always failing every single time like they are now, as well as in some cases erroring out entirely which results in the cache not being uploaded.